Separated the abstract base class EventBackend from the header file of the EventManager. Now Backends can be instantiated without also instatiating the EventManager singleton.
