+++ src/Core/Utility/EventManager.h
@@ -16,6 +16,7 @@
 
 class Event;
 class EventManager;
+class EventBackend;
 
 typedef boost::intrusive_ptr<Event> event_ptr;
 typedef std::vector<event_ptr> EventList;
@@ -30,29 +31,6 @@ table) has one or more inconsistent field-value type pairs.
 class CycInvalidSchemaErr: public CycException {
   public:
     CycInvalidSchemaErr(std::string msg) : CycException(msg) {};
-};
-
-/*!
-An abstract base class for listeners (e.g. output databases) that want
-to receive data generated by the simulation.
-*/
-class EventBackend {
-  public:
-
-    /// Used to pass a list of new/collected events
-    virtual void notify(EventList events) = 0;
-
-    /*!
-    Used to uniquely identify a backend - particularly if there are more
-    than one in a simulation.
-    */
-    virtual std::string name() = 0;
-
-    /*!
-    Used to notify the backend that no more events will be sent (i.e.
-    the simulatoin is over).
-    */
-    virtual void close() = 0;
 };
 
 /*!
